tests/app_django/requirements.txt

# ==== TEST APIs ====

# https://github.com/openskynetwork/opensky-api
pandas
requests
Django
eventlet
gevent_socketio
python-socketio
datetime

